requirements.txt
Flask==2.3.2
flask-cors==3.0.10
torch==2.0.0
torchvision==0.15.0
Pillow==10.0.0
requests==2.31.0
